BestReviewer/RF_Exercise_Before-dev/RF_Exercise_Before-dev/src/main/java/replace/typecode/with/subclass/delegate/Employee.java
package replace.typecode.with.subclass.delegate;

public class Employee {
    private EmployeeType employeeType;
    public Employee(int arg) {
        setEmployeeType(arg);
    }

    public int monthlySalary;
    public int commission;
    public int bonus;
    public int payAmount() {
        return employeeType.payAmount(this);
    }

    public int getEmployeeType() {
        return employeeType.getEmployeeType();
    }

    public void setEmployeeType(int type) {
        this.employeeType = EmployeeType.create(type);
    }
}
